pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "climate-utils"
version = "0.1.0"
description = "Shared utilities for climate data processing and weather file manipulation"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Climate Utils Team"},
]
keywords = ["climate", "weather", "epw", "building", "energy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "python-dateutil>=2.8.0",
    "ladybug-core>=0.44.19",
    "psychrolib>=2.5.0",
]

[project.optional-dependencies]
extended = [
    "psychrolib>=2.5.0",
    "pvlib>=0.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-username/climate-utils"
Repository = "https://github.com/your-username/climate-utils"
Documentation = "https://climate-utils.readthedocs.io"
Issues = "https://github.com/your-username/climate-utils/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/climate_utils"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true